requirements.txt
amqp==2.6.0
asgiref==3.2.9
asn1crypto==1.3.0
bcrypt==3.1.7
beautifulsoup4==4.9.1
billiard==3.6.3.0
celery==4.4.5
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
ConcurrentLogHandler==0.9.1
config==0.5.0.post0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
dj-database-url==0.5.0
dj-email-url==1.0.1
dj-search-url==0.1
Django==3.0.7
django-cache-url==3.1.2
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-configurations==2.2
django-cors-headers==3.4.0
django-filter==2.3.0
django-modeltranslation==0.15
django-suit==0.2.28
django-timezone-field==4.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
dnspython==1.16.0
ephem==3.7.7.1
eventlet==0.25.2
Fabric3==1.14.post1
flake8==3.8.3
future==0.18.2
gevent==20.6.2
greenlet==0.4.16
gunicorn==20.0.4
idna==2.9
itypes==1.2.0
Jinja2==2.11.2
kombu==4.6.10
lxml==4.5.1
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
monotonic==1.5
olefile==0.46
paramiko==2.7.1
phonenumbers==8.12.5
Pillow==7.1.2
psycopg2==2.8.5
psycopg2-binary==2.8.5
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyJWT==1.7.1
PyNaCl==1.4.0
python-crontab==2.5.1
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
vine==1.3.0
zope.event==4.4
zope.interface==5.1.0
